refactor: modularize visit counter logic and API interactions
- Extracted `VisitCounter` component into its own file.
- Moved API interaction logic to a dedicated `visitService`.
- Centralized visit storage handling in `visitStorage` utility.
- Simplified `getServerSideProps` in `index.tsx` by utilizing the new `visitService`.
